gestion de lot entrée en stock 3

+++ src/main/java/com/kobe/warehouse/service/stock/dto/PeremptionStatut.java
@@ -1,3 +1,3 @@
 package com.kobe.warehouse.service.stock.dto;
 
-public record PeremptionStatut(String libelle, int days, int mouths, int years) {}
+public record PeremptionStatut(String libelle, long days, long mouths, int years) {}
